pyproject.toml
[project]
name = "forecasting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pyarrow>=18.0.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "statsmodels>=0.14.5",
    "torch>=2.9.1",
    "tqdm>=4.66.0",
]
